+++ packages/ui-causal-graph-editor/src/shared/rendering/engine.tsx
@@ -27,7 +27,7 @@ import { Status } from '@darajs/ui-utils';
 
 import { CustomLayout, FcoseLayout, GraphLayout } from '@shared/graph-layout';
 import { DragMode } from '@shared/use-drag-mode';
-import { getGroupToNodesMap, getNodeGroup, getNodeToGroupMap } from '@shared/utils';
+import { getGroupToNodesMap, getNodeCategory, getNodeToGroupMap } from '@shared/utils';
 
 import {
     EdgeConstraint,
@@ -418,13 +418,10 @@ export class Engine extends PIXI.utils.EventEmitter<EngineEvents> {
 
             // first create all group nodes so that we have something to connect the edges to
             groupsArray.forEach((group) => {
-                const container = this.groupContainerMap.get(group).groupContainerGfx;
                 const groupNodeAttributes: GroupNode = {
                     id: group,
                     originalMeta: {},
                     variable_type: 'groupNode',
-                    x: container.x,
-                    y: container.y,
                 };
                 // remove all group containers
                 this.dropGroupContainer(group);
@@ -585,6 +582,7 @@ export class Engine extends PIXI.utils.EventEmitter<EngineEvents> {
             // redraw all group containers
             this.createGroupContainers();
 
+            this.debouncedUpdateLayout();
             this.requestRender();
         }
     }
@@ -1307,11 +1305,11 @@ export class Engine extends PIXI.utils.EventEmitter<EngineEvents> {
      * @param attributes node attributes
      */
     private getNodeStyle(node: NodeObject, attributes: SimulationNode): PixiNodeStyle {
-        const group = getNodeGroup(this.graph, attributes.id, attributes['meta.rendering_properties.latent']);
+        const group = getNodeCategory(this.graph, attributes.id, attributes['meta.rendering_properties.latent']);
 
         return {
             color: attributes['meta.rendering_properties.color'],
-            group,
+            category: group,
             highlight_color: attributes['meta.rendering_properties.highlight_color'],
             isEdgeSelected: !!this.selectedEdge,
             isSourceOfNewEdge: this.isCreatingEdge && this.mousedownNodeKey === attributes.id,